i128 -> i64

+++ mods/dr_game/Difficult_Rocket_rs/src/src/data_type.rs
@@ -10,4 +10,4 @@ pub mod dr;
 pub mod math;
 pub mod sr1;
 
-pub type IdType = i128;
+pub type IdType = i64;
